Report on CSF Math
Wrote up my best guesses on what kind of numbers we're going to be
looking at for colonists transported and the mass of the ship and fuel.

You can view Foundating Math.pdf or check out Foundation Math.tex if you
want to see how I made it or make changes.